Album Repository has been added.

+++ backend/src/main/java/org/example/sem_3_eksamen_programmering_lukas/model/Album.java
@@ -6,8 +6,7 @@ import jakarta.persistence.*;
 public class Album {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
+    private int id;
     private String title;
     private String artist;
     private boolean available;
@@ -16,33 +15,42 @@ public class Album {
     @JoinColumn(name = "store_id")
     private Store store;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
-    public void setId(long id) {
+
+    public void setId(int id) {
         this.id = id;
     }
+
     public String getTitle() {
         return title;
     }
+
     public void setTitle(String title) {
         this.title = title;
     }
+
     public String getArtist() {
         return artist;
     }
+
     public void setArtist(String artist) {
         this.artist = artist;
     }
+
     public boolean isAvailable() {
         return available;
     }
+
     public void setAvailable(boolean available) {
         this.available = available;
     }
+
     public Store getStore() {
         return store;
     }
+
     public void setStore(Store store) {
         this.store = store;
     }
